тесты-калькулятор 4

+++ src/test/java/com/skypro/testikurs3kalkulator/CalculatorParametrizedTest.java
@@ -13,43 +13,43 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class CalculatorParametrizedTest {
     private final CalculatorService calculatorService = new CalculatorService();
-    private static final int NUMBER_5 = 5;
-    private static final int NUMBER_6 = 6;
-    private static final int NUMBER_7 = 7;
-    private static final int NUMBER_8 = 8;
-    private static final int ZERO = 0;
+    private static final Integer NUMBER_5 = 5;
+    private static final Integer NUMBER_6 = 6;
+    private static final Integer NUMBER_7 = 7;
+    private static final Integer NUMBER_8 = 8;
+    private static final Integer ZERO = 0;
 //    private int a;
 //    private int b;
 
 
     @ParameterizedTest
     @MethodSource("parametersForTests")
-    public void plusParametrizedTest1(int a, int b) {
-        int result = calculatorService.plus(a, b);
+    public void plusParametrizedTest1(Integer a, Integer b) {
+        Integer result = calculatorService.plus(a, b);
         assertEquals(a + b, result);
     }
 
 
     @ParameterizedTest
     @MethodSource("parametersForTests")
-    public void minusParametrizedTest1(int a, int b) {
-        int result = calculatorService.minus(a, b);
+    public void minusParametrizedTest1(Integer a, Integer b) {
+        Integer result = calculatorService.minus(a, b);
         assertEquals(a - b, result);
     }
 
 
     @ParameterizedTest
     @MethodSource("parametersForTests")
-    public void divisionParametrizedTest1(int a, int b) {
-        int result = (int) calculatorService.division(a, b);
-        assertEquals(a / b, result);
+    public void divisionParametrizedTest1(Integer a, Integer b) {
+        double result = calculatorService.division(a, b);
+        assertEquals((double)a / b, result);
     }
 
 
     @ParameterizedTest
     @MethodSource("parametersForTests")
-    public void multiplicationParametrizedTest1(int a, int b) {
-        int result = (int) calculatorService.multiplication(a, b);
+    public void multiplicationParametrizedTest1(Integer a, Integer b) {
+        Integer result = (int) calculatorService.multiplication(a, b);
         assertEquals(a * b, result);
     }
